+++ src/cli.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 const args = require('yargs-parser')(process.argv.slice(2))
-const fs2json = require('./index')
+const fs2json = require('./index').sync
 
 function main() {